+++ app/models/interest_region.rb
@@ -1,0 +1,4 @@
+class InterestRegion < ApplicationRecord
+  belongs_to :region
+  belongs_to :account
+end